refactor: 필드 분리
